# apps/server/openapi.json
{"openapi":"3.0.3","info":{"title":"Order Wizard API","description":"API for managing Amazon orders with OAuth authentication.\n\n## Authentication Flow\n\n1. Navigate to `/auth/login` to initiate OAuth login (redirects to AWS Cognito)\n2. After successful login, you'll be redirected back with a session cookie\n3. Use the session cookie for authenticated API requests\n4. Call `POST /auth/logout` to end your session","license":{"name":""},"version":"0.1.0"},"paths":{"/auth/me":{"get":{"tags":["Authentication"],"operationId":"current_session","responses":{"200":{"description":"Current session information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionSnapshot"}}}},"401":{"description":"Unauthorized - not logged in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/orders":{"get":{"tags":["Orders"],"operationId":"list_orders","parameters":[{"name":"userId","in":"query","required":false,"schema":{"type":"string","nullable":true}}],"responses":{"200":{"description":"List orders","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}},"400":{"description":"Missing user id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"post":{"tags":["Orders"],"operationId":"create_order","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrder"}}},"required":true},"responses":{"201":{"description":"Order created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}},"400":{"description":"Missing user id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/orders/{id}":{"get":{"tags":["Orders"],"operationId":"get_order","parameters":[{"name":"id","in":"path","description":"Order identifier","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Order detail","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}},"404":{"description":"Order not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["Orders"],"operationId":"delete_order","parameters":[{"name":"id","in":"path","description":"Order identifier","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Order deleted"},"404":{"description":"Order not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"patch":{"tags":["Orders"],"operationId":"update_order","parameters":[{"name":"id","in":"path","description":"Order identifier","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrder"}}},"required":true},"responses":{"204":{"description":"Order updated"},"404":{"description":"Order not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Database error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"CreateOrder":{"type":"object","required":["orderNumber","productName","orderDate","productImage","price","status"],"properties":{"id":{"type":"string","nullable":true},"note":{"type":"string","nullable":true},"orderDate":{"type":"string"},"orderNumber":{"type":"string"},"price":{"type":"string"},"productImage":{"type":"string"},"productName":{"type":"string"},"status":{"$ref":"#/components/schemas/OrderStatus"},"userId":{"type":"string","nullable":true}}},"ErrorResponse":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},"OAuthUser":{"type":"object","required":["id"],"properties":{"email":{"type":"string","nullable":true},"id":{"type":"string"},"name":{"type":"string","nullable":true}}},"Order":{"type":"object","required":["id","userId","orderNumber","productName","orderDate","productImage","price","status"],"properties":{"id":{"type":"string"},"note":{"type":"string","nullable":true},"orderDate":{"type":"string"},"orderNumber":{"type":"string"},"price":{"type":"string"},"productImage":{"type":"string"},"productName":{"type":"string"},"status":{"$ref":"#/components/schemas/OrderStatus"},"userId":{"type":"string"}}},"OrderStatus":{"type":"string","enum":["uncommented","commented","commentRevealed","reimbursed"]},"SessionSnapshot":{"type":"object","required":["user","profile"],"properties":{"expiresAt":{"type":"string","format":"date-time","nullable":true},"profile":{},"user":{"$ref":"#/components/schemas/OAuthUser"}}},"UpdateOrder":{"type":"object","properties":{"note":{"type":"string","nullable":true},"orderDate":{"type":"string","nullable":true},"orderNumber":{"type":"string","nullable":true},"price":{"type":"string","nullable":true},"productImage":{"type":"string","nullable":true},"productName":{"type":"string","nullable":true},"status":{"allOf":[{"$ref":"#/components/schemas/OrderStatus"}],"nullable":true},"userId":{"type":"string","nullable":true}}}}},"tags":[{"name":"Orders","description":"Order management endpoints"},{"name":"Authentication","description":"Authentication session information"}]}